backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
ecdsa==0.19.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.24
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
